feat: complete REST API for cooked schema
Implements all endpoints from the new PostgreSQL DDL: auth, users, ingredients,
recipes (with moods/ingredients/instructions sub-resources), pantry (FIFO deduction),
shopping list (upsert), cook history (kcal snapshot + pantry deduction), saved recipes,
and vocabulary endpoints. JWT userId updated to BIGINT. 16 tests passing.

+++ src/main/java/com/humanworkstream/cooked/security/SecurityUtils.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class SecurityUtils {
 
-    public Integer getCurrentUserId() {
+    public Long getCurrentUserId() {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         if (auth != null && auth.getPrincipal() instanceof UserPrincipal p) {
             return p.userId();
