+++ src/main/java/com/humanworkstream/cooked/security/UserPrincipal.java
@@ -1,4 +1,4 @@
 package com.humanworkstream.cooked.security;
 
-public record UserPrincipal(String email, int userId) {
+public record UserPrincipal(String email, long userId) {
 }
